requirements.txt
flask==3.0.2
requests==2.31.0
gunicorn==22.0.0
python-dotenv==1.0.1
Flask-Caching==2.3.0
Flask-Login==0.5.0
Flask-Bootstrap==3.3.7.1